mfcc_features
This folder contains mfcc features for all the wav files (both for individual wav files and combined in one file). MFCCs were extracted for all the labeled segments of all the wav files. Window size 50ms( ie 800 samples for SR 16000) or  the size of that particular segment whichever is less, and overlap of 50%.